request: send values in Query string instead of POST body
The Google reCaptcha backend apparently expects values
as query string values instead of in the POST body, despite
the documentation's claims, as I detail in issue #1 after
hours of debugging and sweating.

Also adds an example as well as a test to lock in the
discerned behavior. Whenever Google fixes their API,
these tests will fail spectacularly and we'll be able
to switch over easily.

Fixes #1.